+++ lib/geolix/database/loader.ex
@@ -5,11 +5,6 @@ defmodule Geolix.Database.Loader do
 
   use GenServer
 
-  alias Geolix.Adapter.MMDB2.Decoder
-  alias Geolix.Adapter.MMDB2.Metadata
-  alias Geolix.Adapter.MMDB2.Reader
-  alias Geolix.Adapter.MMDB2.Storage
-
   # GenServer lifecycle
 
   @doc """
@@ -23,17 +18,28 @@ defmodule Geolix.Database.Loader do
   def init(databases) do
     init_databases(databases)
 
-    { :ok, databases }
+    state =
+      databases
+      |> Enum.map(&fix_legacy/1)
+      |> Enum.map(fn (database) -> { database[:id], database } end)
+
+    { :ok, state }
   end
 
 
   # GenServer callbacks
 
-  def handle_call({ :set_database, which, filename }, _, state) do
-    case load_database(which, filename) do
-      :ok   -> { :reply, :ok, Keyword.put(state, which, filename) }
+  def handle_call({ :load_database, database }, _, state) do
+    case load_database(database) do
+      :ok   -> { :reply, :ok, Keyword.put(state, database[:id], database) }
       error -> { :reply, error, state }
     end
+  end
+
+  def handle_call({ :set_database, which, filename }, from, state) do
+    database = fix_legacy({ which, filename })
+
+    handle_call({ :load_database, database }, from, state)
   end
 
   def handle_call(:registered, _, state) do
@@ -43,34 +49,59 @@ defmodule Geolix.Database.Loader do
 
   # Internal methods
 
+  defp fix_legacy(database) when is_map(database), do: database
+  defp fix_legacy({ which, filename }) do
+    IO.write :stderr, "The database '#{ inspect which }' is loaded using" <>
+                      " a deprecated tuple definition. Please update to" <>
+                      " the new map based format."
+
+    %{
+      id:      which,
+      adapter: Geolix.Adapter.MMDB2,
+      source:  filename
+    }
+  end
+
+
   defp init_databases([]), do: []
-  defp init_databases([{ which, filename } | databases ]) do
-    load_database(which, filename)
+  defp init_databases([ database | databases ]) do
+    database
+    |> fix_legacy()
+    |> load_database()
+
     init_databases(databases)
   end
 
-  defp load_database(which, { :system, var }) do
-    load_database(which, System.get_env(var))
+  defp load_database(%{ source: { :system, var }} = database) do
+    database
+    |> Map.put(:source, System.get_env(var))
+    |> load_database()
   end
 
-  defp load_database(which, filename) do
-    filename
-    |> Reader.read_database()
-    |> split_data()
-    |> store_data(which)
+  defp load_database(%{ source: source, adapter: adapter } = database) do
+    reader = Module.concat([ adapter, Reader ])
+
+    source
+    |> reader.read_database()
+    |> split_data(database)
+    |> store_data(database)
   end
 
-  defp split_data({ :error, _reason } = error), do: error
-  defp split_data({ data, meta }) do
-    meta           = Decoder.value(meta, 0)
-    meta           = struct(%Metadata{}, meta)
+
+  defp split_data({ :error, _reason } = error, _), do: error
+  defp split_data({ data, meta }, %{ adapter: adapter }) do
+    decoder    = Module.concat([ adapter, Decoder ])
+    metastruct = Module.concat([ adapter, Metadata ])
+
+    meta           = decoder.value(meta, 0)
+    meta           = struct(metastruct, meta)
     record_size    = Map.get(meta, :record_size)
     node_count     = Map.get(meta, :node_count)
     node_byte_size = div(record_size, 4)
     tree_size      = node_count * node_byte_size
 
-    meta = %Metadata{ meta | node_byte_size: node_byte_size }
-    meta = %Metadata{ meta | tree_size:      tree_size }
+    meta = %{ meta | node_byte_size: node_byte_size }
+    meta = %{ meta | tree_size:      tree_size }
 
     tree      = data |> binary_part(0, tree_size)
     data_size = byte_size(data) - byte_size(tree) - 16
@@ -79,11 +110,15 @@ defmodule Geolix.Database.Loader do
     { tree, data, meta }
   end
 
-  defp store_data({ :error, _reason } = error, _which), do: error
-  defp store_data({ tree, data, meta }, which) do
-    Storage.Data.set(which, data)
-    Storage.Metadata.set(which, meta)
-    Storage.Tree.set(which, tree)
+  defp store_data({ :error, _reason } = error, _), do: error
+  defp store_data({ tree, data, meta }, %{ id: id, adapter: adapter}) do
+    storage_data = Module.concat([ adapter, Storage.Data ])
+    storage_meta = Module.concat([ adapter, Storage.Metadata ])
+    storage_tree = Module.concat([ adapter, Storage.Tree ])
+
+    storage_data.set(id, data)
+    storage_meta.set(id, meta)
+    storage_tree.set(id, tree)
 
     :ok
   end